requirements.txt
streamlit==1.42.0
langchain==0.3.18
langchain-community==0.3.17
langchain-huggingface==0.1.2
langchain-groq==0.2.4
faiss-cpu==1.10.0
pypdf==5.3.0
python-dotenv==1.0.1
sentence-transformers==3.4.1
